requirements.txt
matplotlib==3.10.1
mediapipe==0.10.21
numpy==2.2.4
opencv_contrib_python==4.11.0.86
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.1.0
streamlit==1.43.0
scikit_learn==1.6.1
seaborn==0.13.2
timm==1.0.15
torch==2.6.0
torch_ema==0.3
torchvision==0.21.0
tqdm==4.67.1
